merge staging into master branch

+++ src/Component/List/List.js
@@ -1,14 +1,16 @@
-import React, { Component } from "react";
+import React, { Component, Fragment } from "react";
 import { Droppable, Draggable } from "react-beautiful-dnd";
 import CardForm from "../Card/CardForm";
 import Card from "../Card/Card";
+import ListMenu from "./ListMenu";
 
 class List extends Component {
   state = {
     isEdit: false,
     isSubmitted: false,
     showCardForm: false,
-    cardVal: ""
+    cardVal: "",
+    listMenuOpen: false
   };
 
   toggleTitleForm = () => {
@@ -59,7 +61,17 @@ class List extends Component {
         cardVal: ""
       });
     }
+    this.setState({
+      showCardForm: !this.state.showCardForm
+    })
   };
+
+  toggleListMenu = () => {
+    this.setState({
+      listMenuOpen: !this.state.listMenuOpen
+    })
+  }
+
 
   // WILL USE FUNCTION INSIDE OF MODAL TO DELETE CARD
   // deleteCard = id => {
@@ -70,7 +82,7 @@ class List extends Component {
   // };
 
   render() {
-    const { isEdit, isSubmitted, showCardForm, cardVal } = this.state;
+    const { isEdit, isSubmitted, showCardForm, cardVal, listMenuOpen } = this.state;
     const { id, title } = this.props.list;
     const { handleTitleChange, cardList, addCardDescription } = this.props;
     return (
@@ -96,9 +108,23 @@ class List extends Component {
                     />
                     {// if editing list title, no need to show "Add List" button
                       !isEdit && <button>Add List</button>}
+                      {
+                      isEdit && <button>Edit List</button>}
                   </form>
                 ) : (
-                    <h3 onClick={this.toggleTitleForm}>{title}</h3>
+                    <Fragment>
+                      <h3 onClick={this.toggleTitleForm}>{title}</h3>
+                      <button className="open-list-menu-btn" onClick={this.toggleListMenu}><i className="fas fa-ellipsis-h fa-sm"></i></button>
+                      {
+                        listMenuOpen &&
+                        <ListMenu
+                          toggleListMenu={this.toggleListMenu}
+                          deleteList={this.props.deleteList}
+                          listId={id}
+                        />
+                      }
+
+                    </Fragment>
                   )}
             </div>
             {provided.placeholder}
@@ -153,15 +179,19 @@ class List extends Component {
                   + <span>Add a card...</span>
                 </p>
               )}
-            <button
+            {/* <button
               className="btn btn--delete-list"
               onClick={e => {
                 e.preventDefault();
-                this.props.deleteList(id);
+                if(window.confirm('Are you sure you want to delete ' + title + '?')) {
+                  alert("List " + title + " deleted.");
+                  this.props.deleteList(id);
+                } else {}
+                // this.props.deleteList(id);
               }}
             >
               Delete List
-        </button>
+        </button> */}
           </div>)}
 
       </Draggable>
